Split sources to modules
