+++ src/dictionary/main.js
@@ -1,0 +1,2438 @@
+/* YoptaScript
+ * Массив без лишних экранирований
+ * Для перевода из js в ys
+ */
+module.exports = [
+  [
+    "includes",
+    "лучшеНетВлагалищаЧемОчкоТоварища"
+  ],
+  [
+    "createRadialGradient",
+    "намутитьПоКругуМазнюПодливой"
+  ],
+  [
+    "createLinearGradient",
+    "намутитьЧоткуюМазнюПодливой"
+  ],
+  [
+    "resizeTo",
+    "распидораситьОтносительно"
+  ],
+  [
+    "fromCodePoint",
+    "хуйняИзЭтоТуданахНутыпоэл"
+  ],
+  [
+    "createTextNode",
+    "намутитьМалявуГовнодскую"
+  ],
+  [
+    "enableStyleSheetsForSet",
+    "намутитьСтруйкуДляХабара"
+  ],
+  [
+    "ondragenter",
+    "покаДёргалкаНарисовалась"
+  ],
+  [
+    "ondragenter",
+    "покаДергалкаНарисовалась"
+  ],
+  [
+    "imul",
+    "петухПетухаВидитИздалека"
+  ],
+  [
+    "getElementsByClassName",
+    "вычислитьТерпилПоКлассу"
+  ],
+  [
+    "vlinkColor",
+    "когдаОткинулсяПослеЗоны"
+  ],
+  [
+    "onpointermove",
+    "покаХожуТырколйПоРайону"
+  ],
+  [
+    "getElementsByName",
+    "вычислитьТерпилПоИмени"
+  ],
+  [
+    "onpointerenter",
+    "покаСрелкаНарисовалась"
+  ],
+  [
+    "oninstall",
+    "покаХуйСтоитКакКолонна"
+  ],
+  [
+    "isPointInStroke",
+    "естьЛиКуполаНаНаколках"
+  ],
+  [
+    "getElementsByTagName",
+    "вычислитьТерпилПоТегу"
+  ],
+  [
+    "getElementById",
+    "вычислитьЛохаПоНомеру"
+  ],
+  [
+    "querySelectorAll",
+    "хулиТутВсеВыёбываются"
+  ],
+  [
+    "querySelectorAll",
+    "хулиТутВсеВыебываются"
+  ],
+  [
+    "pointerLockElement",
+    "тырколкуНаАнусТерпилы"
+  ],
+  [
+    "getComputedStyle",
+    "нассыМнеВалиоСтруйкой"
+  ],
+  [
+    "imageSmoothingEnabled",
+    "включитьРазмытиеЕбала"
+  ],
+  [
+    "scrollPathIntoView",
+    "колеситьПоДорогеНахуй"
+  ],
+  [
+    "setLineDash",
+    "заебенитьПятнистыйХуй"
+  ],
+  [
+    "onreadystatechange",
+    "опаГотовЙоптЧозанахуй"
+  ],
+  [
+    "createNodeIterator",
+    "намутитьГовнодыратор"
+  ],
+  [
+    "lastElementChild",
+    "последнийПездюкШняги"
+  ],
+  [
+    "onafterscriptexecute",
+    "послеВыполненияЙопты"
+  ],
+  [
+    "scrollbars",
+    "колеситьПоПивларькам"
+  ],
+  [
+    "resizeBy",
+    "распидораситьПоХуйне"
+  ],
+  [
+    "sizeToContent",
+    "ухтыжёптыжбляПодгони"
+  ],
+  [
+    "sizeToContent",
+    "ухтыжептыжбляПодгони"
+  ],
+  [
+    "fontcolor",
+    "говномПоСтенеКрасиво"
+  ],
+  [
+    "removeHitRegion",
+    "избавитьсяОтПроблемы"
+  ],
+  [
+    "getResponseHeader",
+    "дайКепарикПолосатого"
+  ],
+  [
+    "acosh",
+    "агопосинусКупчинский"
+  ],
+  [
+    "atanh",
+    "агопангенсКупчинский"
+  ],
+  [
+    "createDocumentFragment",
+    "намутитьКусокМалявы"
+  ],
+  [
+    "onselectionchange",
+    "покаДваСтулаМахнуть"
+  ],
+  [
+    "onfullscreenchange",
+    "покаЕбальникПоказал"
+  ],
+  [
+    "ondblclick",
+    "какПырнуДваждыНахуй"
+  ],
+  [
+    "ondragend",
+    "покаХарэОчкоДёргать"
+  ],
+  [
+    "ondragend",
+    "покаХарэОчкоДергать"
+  ],
+  [
+    "oninvalid",
+    "гыйбатьИнвалидНахуй"
+  ],
+  [
+    "onmouseout",
+    "покаТырколкаСъебала"
+  ],
+  [
+    "onmouseover",
+    "покаТырколкаПодошла"
+  ],
+  [
+    "onpointercancel",
+    "покаСтрелаТруханула"
+  ],
+  [
+    "onpopstate",
+    "покаИсторияМаляется"
+  ],
+  [
+    "isDefaultNamespace",
+    "деткаТыПростоКосмос"
+  ],
+  [
+    "codePointAt",
+    "хуйняНутыпоэлОткуда"
+  ],
+  [
+    "isPointInPath",
+    "естьЛиКуполаНаГруди"
+  ],
+  [
+    "XMLHttpRequestEventTarget",
+    "запросСоШнягойПоЩам"
+  ],
+  [
+    "cosh",
+    "гопосиносКолпинский"
+  ],
+  [
+    "tanh",
+    "гопангенсКолпинский"
+  ],
+  [
+    "ondrop",
+    "опаМабилкаЁбнулась"
+  ],
+  [
+    "ondrop",
+    "опаМабилкаЕбнулась"
+  ],
+  [
+    "oninput",
+    "покаЭйтыэтоПишибля"
+  ],
+  [
+    "ontouchmove",
+    "опаДвигайОтСюдаЛещ"
+  ],
+  [
+    "scrollMaxX",
+    "колеситьПоГлавЭксу"
+  ],
+  [
+    "scrollMaxY",
+    "колеситьПоГлавУгам"
+  ],
+  [
+    "compareDocumentPosition",
+    "сравниСтатусМалявы"
+  ],
+  [
+    "substring",
+    "спиздитьМеждуСтрок"
+  ],
+  [
+    "trimRight",
+    "вырезатьОчкоСправа"
+  ],
+  [
+    "fontsize",
+    "говномПоСтенеСочно"
+  ],
+  [
+    "createImageData",
+    "намутитьМазнюЙопта"
+  ],
+  [
+    "setTransform",
+    "перекоситьЕбальник"
+  ],
+  [
+    "parseFloat",
+    "шнырятьПоПлавникам"
+  ],
+  [
+    "dirxml",
+    "йбатьБуратиноНахуй"
+  ],
+  [
+    "readystatechange",
+    "готовЙоптЧозанахуй"
+  ],
+  [
+    "LOG10E",
+    "СЛОЖНЫЙ_ГОПОРИФМ10"
+  ],
+  [
+    "SQRT1_2",
+    "сквиртНаПолшишечки"
+  ],
+  [
+    "asinh",
+    "агопинусКупчинский"
+  ],
+  [
+    "queryCommandEnabled",
+    "хулиЧикаДоступная"
+  ],
+  [
+    "queryCommandSupported",
+    "хулиЧикаБезАйфона"
+  ],
+  [
+    "firstElementChild",
+    "первыйПездюкШняги"
+  ],
+  [
+    "childElementCount",
+    "моиШняжныеПездюки"
+  ],
+  [
+    "onbeforescriptexecute",
+    "доВыполненияЙопты"
+  ],
+  [
+    "ondragover",
+    "покаДёрнулПодошла"
+  ],
+  [
+    "ondragover",
+    "покаДернулПодошла"
+  ],
+  [
+    "onpointerover",
+    "покаСтрелаПодошла"
+  ],
+  [
+    "onpointerout",
+    "покаСтрелаСъебала"
+  ],
+  [
+    "onpointerleave",
+    "покаСтрелаСлиняла"
+  ],
+  [
+    "ontouchstart",
+    "опаЩаЛещаПоЩамДам"
+  ],
+  [
+    "dispatchEvent",
+    "послатьНахуйШнягу"
+  ],
+  [
+    "scrollByPages",
+    "колеситьНаМалявах"
+  ],
+  [
+    "updateCommands",
+    "новыйАйфонДляЧики"
+  ],
+  [
+    "onbeforeunload",
+    "покаСемкиКрутятся"
+  ],
+  [
+    "onhashchange",
+    "покаШнягаИзменяет"
+  ],
+  [
+    "onpagehide",
+    "покаКсиваНаМалине"
+  ],
+  [
+    "hasChildNodes",
+    "батяИмеетПездюков"
+  ],
+  [
+    "insertBefore",
+    "вставитьПездюкаДо"
+  ],
+  [
+    "substr",
+    "спиздитьМеждуБукв"
+  ],
+  [
+    "trimLeft",
+    "вырезатьОчкоСлева"
+  ],
+  [
+    "quadraticCurveTo",
+    "криваяЗавелаНахуй"
+  ],
+  [
+    "resetTransform",
+    "поправитьЕбальник"
+  ],
+  [
+    "parseInt",
+    "шнырятьПоКарманам"
+  ],
+  [
+    "timeout",
+    "длительностьСрока"
+  ],
+  [
+    "loadstart",
+    "началТаскатьСемки"
+  ],
+  [
+    "progress",
+    "сколькоСемокДонёс"
+  ],
+  [
+    "progress",
+    "сколькоСемокДонес"
+  ],
+  [
+    "LOG2E",
+    "СЛОЖНЫЙ_ГОПОРИФМ2"
+  ],
+  [
+    "clz32",
+    "поводырь32петухов"
+  ],
+  [
+    "log1p",
+    "чистыйГопорифмПо1"
+  ],
+  [
+    "sinh",
+    "гопинусКолпинский"
+  ],
+  [
+    "createTreeWalker",
+    "намутитьБуратино"
+  ],
+  [
+    "getAnimations",
+    "вычислитьДвижуху"
+  ],
+  [
+    "releaseCapture",
+    "зафотатьХуякХуяк"
+  ],
+  [
+    "querySelector",
+    "хулиВыёбываешься"
+  ],
+  [
+    "querySelector",
+    "хулиВыебываешься"
+  ],
+  [
+    "queryCommandState",
+    "хулиЧикаОтдыхает"
+  ],
+  [
+    "contentType",
+    "ухтыжёптыжТипчик"
+  ],
+  [
+    "contentType",
+    "ухтыжептыжТипчик"
+  ],
+  [
+    "alinkColor",
+    "петухЗоныКрасиво"
+  ],
+  [
+    "lastModified",
+    "когдаПетухомСтал"
+  ],
+  [
+    "ondragexit",
+    "анусСебеДёрниПёс"
+  ],
+  [
+    "ondragexit",
+    "анусСебеДерниПес"
+  ],
+  [
+    "ondragleave",
+    "покаДергунСлинял"
+  ],
+  [
+    "onmouseleave",
+    "съебалсяИзРайона"
+  ],
+  [
+    "onpointerdown",
+    "покаТыркнулДауна"
+  ],
+  [
+    "onprogress",
+    "покаМатаетсяСрок"
+  ],
+  [
+    "ontimeupdate",
+    "опаНуЭтоКогдаЭто"
+  ],
+  [
+    "addEventListener",
+    "добавитьВертухай"
+  ],
+  [
+    "setResizable",
+    "датьПопидорасить"
+  ],
+  [
+    "onunload",
+    "опаСемкиКрутятся"
+  ],
+  [
+    "baseURIObject",
+    "мразотыНаХатеКрч"
+  ],
+  [
+    "ownerDocument",
+    "главныйАвторитет"
+  ],
+  [
+    "textContent",
+    "ухтыжёптыжМалява"
+  ],
+  [
+    "textContent",
+    "ухтыжептыжМалява"
+  ],
+  [
+    "contains",
+    "яТвойОтецЕбуОвец"
+  ],
+  [
+    "lastIndexOf",
+    "последняяОтсидка"
+  ],
+  [
+    "toLocaleLowerCase",
+    "поРайонуНеКапсом"
+  ],
+  [
+    "shadowBlur",
+    "наняРазмытьЕбало"
+  ],
+  [
+    "isSafeInteger",
+    "этоОхуеннаяХуйня"
+  ],
+  [
+    "toLocaleString",
+    "кПацанамНаРайоне"
+  ],
+  [
+    "groupEnd",
+    "съебатьсяИзБанды"
+  ],
+  [
+    "reverse",
+    "шухильмеМухильме"
+  ],
+  [
+    "elementFromPoint",
+    "терпилаИзПараши"
+  ],
+  [
+    "elementsFromPoint",
+    "терпилыИзПараши"
+  ],
+  [
+    "getSelection",
+    "сестьНаДваСтула"
+  ],
+  [
+    "onabort",
+    "когдаУронилМыло"
+  ],
+  [
+    "onkeypress",
+    "гыйбатьВмялЛеща"
+  ],
+  [
+    "onpointerup",
+    "покаТыркнулАута"
+  ],
+  [
+    "ontouchcancel",
+    "покаЛещТруханул"
+  ],
+  [
+    "returnValue",
+    "ответитьЗаВалио"
+  ],
+  [
+    "scrollByLines",
+    "колеситьНаЛинии"
+  ],
+  [
+    "scrollTo",
+    "колеситьНахуйНа"
+  ],
+  [
+    "onpageshow",
+    "опаКсивуПоказал"
+  ],
+  [
+    "lastChild",
+    "последнийПездюк"
+  ],
+  [
+    "nextSibling",
+    "следующийПездюк"
+  ],
+  [
+    "nodeName",
+    "погонялоПездюка"
+  ],
+  [
+    "parentElement",
+    "братишкаЭлемент"
+  ],
+  [
+    "appendChild",
+    "заделатьПездюка"
+  ],
+  [
+    "fillStyle",
+    "кончитьСтруйкой"
+  ],
+  [
+    "miterLimit",
+    "скрестимСтруйки"
+  ],
+  [
+    "addHitRegion",
+    "создатьПроблемы"
+  ],
+  [
+    "createPattern",
+    "намутитьТипчика"
+  ],
+  [
+    "getImageData",
+    "чоТутНамалёвано"
+  ],
+  [
+    "getImageData",
+    "чоТутНамалевано"
+  ],
+  [
+    "profileEnd",
+    "вЛичкуПрописали"
+  ],
+  [
+    "onwheel",
+    "какНаХуюВертел"
+  ],
+  [
+    "ondragstart",
+    "покаДёрнулКмон"
+  ],
+  [
+    "ondragstart",
+    "покаДернулКмон"
+  ],
+  [
+    "onkeydown",
+    "гыйбатьЛещДаун"
+  ],
+  [
+    "onmousedown",
+    "всунулНаРайоне"
+  ],
+  [
+    "onmousewheel",
+    "вертелНаРайоне"
+  ],
+  [
+    "onseeking",
+    "когдаОбоссался"
+  ],
+  [
+    "onstalled",
+    "опаНефартануло"
+  ],
+  [
+    "screenX",
+    "всёЕбалоПоЭксу"
+  ],
+  [
+    "screenX",
+    "всеЕбалоПоЭксу"
+  ],
+  [
+    "screenY",
+    "всёЕбалоПоУгам"
+  ],
+  [
+    "screenY",
+    "всеЕбалоПоУгам"
+  ],
+  [
+    "scrollX",
+    "колеситьПоЭксу"
+  ],
+  [
+    "scrollY",
+    "колеситьПоУгам"
+  ],
+  [
+    "ondevicelight",
+    "покаХуйДлинный"
+  ],
+  [
+    "onpaint",
+    "опаНарисовался"
+  ],
+  [
+    "nodeType",
+    "типичныйПездюк"
+  ],
+  [
+    "padStart",
+    "начальныйПадик"
+  ],
+  [
+    "toLocaleUpperCase",
+    "поРайонуКапсом"
+  ],
+  [
+    "shadowOffsetX",
+    "наняХатаПоЭксу"
+  ],
+  [
+    "shadowOffsetY",
+    "наняХатаПоУгам"
+  ],
+  [
+    "textAlign",
+    "выровнитьБазар"
+  ],
+  [
+    "textBaseline",
+    "поставитьБазар"
+  ],
+  [
+    "fillRect",
+    "обкончатьДоску"
+  ],
+  [
+    "fillText",
+    "обкончатьБуквы"
+  ],
+  [
+    "MAX_VALUE",
+    "ОХУЕННОЕ_ВАЛИО"
+  ],
+  [
+    "toExponential",
+    "наХуекспоненту"
+  ],
+  [
+    "trace",
+    "складЧмошников"
+  ],
+  [
+    "readyState",
+    "готовностьЙопт"
+  ],
+  [
+    "entries",
+    "вычислитьЛохов"
+  ],
+  [
+    "expm1",
+    "топГопспонента"
+  ],
+  [
+    "fround",
+    "мелочьТожеГони"
+  ],
+  [
+    "ignoreCase",
+    "игнорщикЕбаный"
+  ],
+  [
+    "sticky",
+    "петухОпущенный"
+  ],
+  [
+    "lastIndex",
+    "доКонцаОтсидки"
+  ],
+  [
+    "test",
+    "ответыБудутЭээ"
+  ],
+  [
+    "captureEvents",
+    "зафотатьШняги"
+  ],
+  [
+    "createEvent",
+    "намутитьШнягу"
+  ],
+  [
+    "createRange",
+    "намутитьОпапа"
+  ],
+  [
+    "createExpression",
+    "намутитьБазар"
+  ],
+  [
+    "execCommand",
+    "идиРаботайБля"
+  ],
+  [
+    "hasFocus",
+    "имеетЧеткость"
+  ],
+  [
+    "queryCommandIndeterm",
+    "хулиЧикаОйВсё"
+  ],
+  [
+    "queryCommandIndeterm",
+    "хулиЧикаОйВсе"
+  ],
+  [
+    "queryCommandValue",
+    "хулиЧикаВалио"
+  ],
+  [
+    "inputEncoding",
+    "эйтыэтоПиздиш"
+  ],
+  [
+    "scrollingElement",
+    "намазиТерпила"
+  ],
+  [
+    "activeElement",
+    "активнаяШняга"
+  ],
+  [
+    "defaultView",
+    "моргалаВыколю"
+  ],
+  [
+    "ondrag",
+    "опаОчкоДёрнул"
+  ],
+  [
+    "ondrag",
+    "опаОчкоДернул"
+  ],
+  [
+    "onkeyup",
+    "гыйбатьЛещАут"
+  ],
+  [
+    "onloadstart",
+    "покаНесуСемки"
+  ],
+  [
+    "onmouseup",
+    "вынулНаРайоне"
+  ],
+  [
+    "onsuspend",
+    "опаПодфартило"
+  ],
+  [
+    "ontouchend",
+    "покаЛещКончил"
+  ],
+  [
+    "onwaiting",
+    "покаМотаюСрок"
+  ],
+  [
+    "frameElement",
+    "кадрОпущенный"
+  ],
+  [
+    "innerHeight",
+    "внутриДлинный"
+  ],
+  [
+    "outerHeight",
+    "вокругДлинные"
+  ],
+  [
+    "sessionStorage",
+    "хабрИзОтсидки"
+  ],
+  [
+    "postMessage",
+    "намутитьКсиву"
+  ],
+  [
+    "prompt",
+    "поясниЗаБазар"
+  ],
+  [
+    "removeEventListener",
+    "урытьВертухая"
+  ],
+  [
+    "getRootNode",
+    "дайБатеГовнод"
+  ],
+  [
+    "isEqualNode",
+    "эквалиоГовнод"
+  ],
+  [
+    "removeChild",
+    "уебатьПездюка"
+  ],
+  [
+    "localeCompare",
+    "сравнитьГовор"
+  ],
+  [
+    "padEnd",
+    "залупныйПадик"
+  ],
+  [
+    "repeat",
+    "непоэлПовтори"
+  ],
+  [
+    "slice",
+    "поделитьСемки"
+  ],
+  [
+    "startsWith",
+    "начатьЗалупку"
+  ],
+  [
+    "font",
+    "говномПоСтене"
+  ],
+  [
+    "globalAlpha",
+    "главныйАльфач"
+  ],
+  [
+    "clearHitRegions",
+    "убитьПроблему"
+  ],
+  [
+    "measureText",
+    "вместитьБазар"
+  ],
+  [
+    "restore",
+    "выздоравливай"
+  ],
+  [
+    "MAX_SAFE_INTEGER",
+    "ЛУЧШИЙ_ГАНДОН"
+  ],
+  [
+    "MIN_SAFE_INTEGER",
+    "ХУЁВЫЙ_ГАНДОН"
+  ],
+  [
+    "MIN_SAFE_INTEGER",
+    "ХУЕВЫЙ_ГАНДОН"
+  ],
+  [
+    "groupCollapsed",
+    "свернутьБанду"
+  ],
+  [
+    "EventTarget",
+    "очкоНаПрицеле"
+  ],
+  [
+    "reduceRight",
+    "редискаПравая"
+  ],
+  [
+    "SQRT2",
+    "двойнойСквирт"
+  ],
+  [
+    "atan2",
+    "агопангенсПо2"
+  ],
+  [
+    "hypot",
+    "вКореньЗыришь"
+  ],
+  [
+    "RegExp",
+    "фильтруйБазар"
+  ],
+  [
+    "global",
+    "глобалкаЙопта"
+  ],
+  [
+    "source",
+    "чоТыБляСказал"
+  ],
+  [
+    "createAttribute",
+    "намутитьЯжку"
+  ],
+  [
+    "createTouch",
+    "намутитьЛеща"
+  ],
+  [
+    "documentElement",
+    "ксиваТерпилы"
+  ],
+  [
+    "bgColor",
+    "охуеннаяЖопа"
+  ],
+  [
+    "oncancel",
+    "покаТруханул"
+  ],
+  [
+    "onchange",
+    "опаЧозанахуй"
+  ],
+  [
+    "oncontextmenu",
+    "какПоЛбуЁбну"
+  ],
+  [
+    "oncontextmenu",
+    "какПоЛбуЕбну"
+  ],
+  [
+    "onmouseenter",
+    "вошёлНаРайон"
+  ],
+  [
+    "onmouseenter",
+    "вошелНаРайон"
+  ],
+  [
+    "onmousemove",
+    "хожуПоРайону"
+  ],
+  [
+    "onseeked",
+    "когдаОбоссал"
+  ],
+  [
+    "onsubmit",
+    "опаХуйВГовне"
+  ],
+  [
+    "onvolumechange",
+    "покаТишеБудь"
+  ],
+  [
+    "console",
+    "красноглазое"
+  ],
+  [
+    "dialogArguments",
+    "тыэтаТавоэта"
+  ],
+  [
+    "innerWidth",
+    "внутриЖирный"
+  ],
+  [
+    "outerWidth",
+    "вокругЖирные"
+  ],
+  [
+    "pageXOffset",
+    "статьяПоЭксу"
+  ],
+  [
+    "pageYOffset",
+    "статьяПоУгам"
+  ],
+  [
+    "blur",
+    "размытьЕбало"
+  ],
+  [
+    "clearInterval",
+    "отсидетьСизо"
+  ],
+  [
+    "clearTimeout",
+    "отсидетьСрок"
+  ],
+  [
+    "moveTo",
+    "нахуйЭтоТуда"
+  ],
+  [
+    "setInterval",
+    "посетитьСизо"
+  ],
+  [
+    "setTimeout",
+    "получитьСрок"
+  ],
+  [
+    "onload",
+    "опаСемкиНесу"
+  ],
+  [
+    "childNodes",
+    "пездюкГовнод"
+  ],
+  [
+    "firstChild",
+    "первыйПездюк"
+  ],
+  [
+    "nodeValue",
+    "валиоПездюка"
+  ],
+  [
+    "previousSibling",
+    "старыйПездюк"
+  ],
+  [
+    "replaceChild",
+    "сделатьАборт"
+  ],
+  [
+    "replace",
+    "пивасПодмени"
+  ],
+  [
+    "trim",
+    "вырезатьОчко"
+  ],
+  [
+    "currentTransform",
+    "этотЕбальник"
+  ],
+  [
+    "lineCap",
+    "выбратьКонец"
+  ],
+  [
+    "lineDashOffset",
+    "пятнистыйХуй"
+  ],
+  [
+    "lineJoin",
+    "формаШишкана"
+  ],
+  [
+    "strokeStyle",
+    "стильНаколок"
+  ],
+  [
+    "putImageData",
+    "намалюйЧоТут"
+  ],
+  [
+    "rotate",
+    "вертетьНаХую"
+  ],
+  [
+    "strokeText",
+    "текстНаколки"
+  ],
+  [
+    "translate",
+    "дисюдаПиксел"
+  ],
+  [
+    "MIN_VALUE",
+    "ХУЁВОЕ_ВАЛИО"
+  ],
+  [
+    "MIN_VALUE",
+    "ХУЕВОЕ_ВАЛИО"
+  ],
+  [
+    "POSITIVE_INFINITY",
+    "ОХУЕТЬ_ДОХУЯ"
+  ],
+  [
+    "assert",
+    "найтиСтукача"
+  ],
+  [
+    "copyWithin",
+    "вынестиГовно"
+  ],
+  [
+    "forEach",
+    "пероПодРебро"
+  ],
+  [
+    "splice",
+    "въебатьГовна"
+  ],
+  [
+    "log10",
+    "гопорифмПо10"
+  ],
+  [
+    "input",
+    "тыЭтоПишибля"
+  ],
+  [
+    "flags",
+    "флагМнеВанус"
+  ],
+  [
+    "exec",
+    "работайМразь"
+  ],
+  [
+    "else",
+    "иливжопураз"
+  ],
+  [
+    "if",
+    "вилкойвглаз"
+  ],
+  [
+    "characterSet",
+    "слышТыЧоЁба"
+  ],
+  [
+    "characterSet",
+    "слышТыЧоЕба"
+  ],
+  [
+    "documentURI",
+    "ксиваНаХате"
+  ],
+  [
+    "domConfig",
+    "чоткоДерзко"
+  ],
+  [
+    "visibilityState",
+    "мутныйСюжет"
+  ],
+  [
+    "designMode",
+    "хуйРисуйМод"
+  ],
+  [
+    "linkColor",
+    "зонаКрасиво"
+  ],
+  [
+    "referrer",
+    "корешСтарый"
+  ],
+  [
+    "onscroll",
+    "покаКолесим"
+  ],
+  [
+    "onselect",
+    "опаДваСтула"
+  ],
+  [
+    "location",
+    "белыйЛебедь"
+  ],
+  [
+    "sidebar",
+    "стенкаЙбать"
+  ],
+  [
+    "cancelIdleCallback",
+    "харэПиздеть"
+  ],
+  [
+    "maximize",
+    "распидорась"
+  ],
+  [
+    "moveBy",
+    "щаТяПодвину"
+  ],
+  [
+    "onoffline",
+    "покаОффнусь"
+  ],
+  [
+    "charCodeAt",
+    "обаЁбаХуйня"
+  ],
+  [
+    "charCodeAt",
+    "обаЕбаХуйня"
+  ],
+  [
+    "split",
+    "поделитьЯгу"
+  ],
+  [
+    "italics",
+    "понаехавший"
+  ],
+  [
+    "shadowColor",
+    "наняКрасиво"
+  ],
+  [
+    "getContext",
+    "снятьСкальп"
+  ],
+  [
+    "NEGATIVE_INFINITY",
+    "НИХУЯ_ДОХУЯ"
+  ],
+  [
+    "isFinite",
+    "оноКонченое"
+  ],
+  [
+    "count",
+    "которыйСрок"
+  ],
+  [
+    "timeStamp",
+    "началоСрока"
+  ],
+  [
+    "XMLHttpRequest",
+    "запросПоЩам"
+  ],
+  [
+    "shift",
+    "первыйБачок"
+  ],
+  [
+    "some",
+    "нарываешься"
+  ],
+  [
+    "E",
+    "ГОПСПАНЕНТА"
+  ],
+  [
+    "exp",
+    "гопспанента"
+  ],
+  [
+    "log2",
+    "гопорифмПо2"
+  ],
+  [
+    "Abstract",
+    "говнойбать"
+  ],
+  [
+    "importNode",
+    "влабазУзел"
+  ],
+  [
+    "registerElement",
+    "зашитьДело"
+  ],
+  [
+    "domain",
+    "домойБлядь"
+  ],
+  [
+    "oncut",
+    "какВысунул"
+  ],
+  [
+    "onpaste",
+    "какВставил"
+  ],
+  [
+    "onblur",
+    "опаНичотка"
+  ],
+  [
+    "onerror",
+    "наПапандос"
+  ],
+  [
+    "onreset",
+    "покаПравим"
+  ],
+  [
+    "fullScreen",
+    "воВсёЕбало"
+  ],
+  [
+    "fullScreen",
+    "воВсеЕбало"
+  ],
+  [
+    "opener",
+    "открывашка"
+  ],
+  [
+    "performance",
+    "сестьНахуй"
+  ],
+  [
+    "top",
+    "КрышаЙбать"
+  ],
+  [
+    "getAttention",
+    "посвистеть"
+  ],
+  [
+    "matchMedia",
+    "феняНаШару"
+  ],
+  [
+    "openDialog",
+    "побазарить"
+  ],
+  [
+    "scrollBy",
+    "колеситьНа"
+  ],
+  [
+    "ononline",
+    "опаТутачки"
+  ],
+  [
+    "cloneNode",
+    "клонГовнод"
+  ],
+  [
+    "prototype",
+    "проточелик"
+  ],
+  [
+    "fromCharCode",
+    "хуйняИзЁба"
+  ],
+  [
+    "fromCharCode",
+    "хуйняИзЕба"
+  ],
+  [
+    "concat",
+    "заебеньВсе"
+  ],
+  [
+    "toString",
+    "поПацански"
+  ],
+  [
+    "big",
+    "большойЁпт"
+  ],
+  [
+    "big",
+    "большойЕпт"
+  ],
+  [
+    "lineWidth",
+    "толщинаХуя"
+  ],
+  [
+    "bezierCurveTo",
+    "кривоНахуй"
+  ],
+  [
+    "closePath",
+    "тупикНахуй"
+  ],
+  [
+    "getLineDash",
+    "сококПятен"
+  ],
+  [
+    "lineTo",
+    "прямоНахуй"
+  ],
+  [
+    "transform",
+    "перекосить"
+  ],
+  [
+    "timeEnd",
+    "конецСрока"
+  ],
+  [
+    "warn",
+    "тыЭтоНуЭто"
+  ],
+  [
+    "withCredentials",
+    "зашкварить"
+  ],
+  [
+    "abort",
+    "уронилМыло"
+  ],
+  [
+    "load",
+    "нестиСемки"
+  ],
+  [
+    "loadend",
+    "семкиДонёс"
+  ],
+  [
+    "loadend",
+    "семкиДонес"
+  ],
+  [
+    "from",
+    "спиздитьИз"
+  ],
+  [
+    "isArray",
+    "этоПомойка"
+  ],
+  [
+    "findIndex",
+    "найдиБомжа"
+  ],
+  [
+    "map",
+    "засратьВсё"
+  ],
+  [
+    "map",
+    "засратьВсе"
+  ],
+  [
+    "unshift",
+    "верниБачок"
+  ],
+  [
+    "LN10",
+    "ГОПОРИФМ10"
+  ],
+  [
+    "acos",
+    "агопосинус"
+  ],
+  [
+    "atan",
+    "агопангенс"
+  ],
+  [
+    "cbrt",
+    "кубоСквирт"
+  ],
+  [
+    "trunc",
+    "верниЧирик"
+  ],
+  [
+    "multiline",
+    "стулБезПик"
+  ],
+  [
+    "unicode",
+    "хуйняНахуй"
+  ],
+  [
+    "switch",
+    "пошерстим"
+  ],
+  [
+    "typeof",
+    "чезажижан"
+  ],
+  [
+    "static",
+    "попонятия"
+  ],
+  [
+    "writeln",
+    "малявагоп"
+  ],
+  [
+    "readyState",
+    "газуемБля"
+  ],
+  [
+    "oncopy",
+    "какВсунул"
+  ],
+  [
+    "onpause",
+    "покаСтопэ"
+  ],
+  [
+    "onplaying",
+    "покаЖивой"
+  ],
+  [
+    "crypto",
+    "пиздишбля"
+  ],
+  [
+    "devicePixelRatio",
+    "типАйфона"
+  ],
+  [
+    "length",
+    "писькомер"
+  ],
+  [
+    "navigator",
+    "главпетух"
+  ],
+  [
+    "minimize",
+    "спидорась"
+  ],
+  [
+    "baseURI",
+    "наХатеТип"
+  ],
+  [
+    "endsWith",
+    "отЗалупки"
+  ],
+  [
+    "search",
+    "семкиЕсть"
+  ],
+  [
+    "toUpperCase",
+    "капсомБля"
+  ],
+  [
+    "bold",
+    "жирныйЁпт"
+  ],
+  [
+    "bold",
+    "жирныйЕпт"
+  ],
+  [
+    "strike",
+    "въебиОчко"
+  ],
+  [
+    "clearRect",
+    "урытьШкаф"
+  ],
+  [
+    "fill",
+    "обкончать"
+  ],
+  [
+    "save",
+    "схоронить"
+  ],
+  [
+    "EPSILON",
+    "ХУЕПСИЛОН"
+  ],
+  [
+    "toPrecision",
+    "кЧоткости"
+  ],
+  [
+    "upload",
+    "заебенить"
+  ],
+  [
+    "send",
+    "всёПиздуй"
+  ],
+  [
+    "send",
+    "всеПиздуй"
+  ],
+  [
+    "join",
+    "вписаться"
+  ],
+  [
+    "LN2",
+    "ГОПОРИФМ2"
+  ],
+  [
+    "abs",
+    "абсолютли"
+  ],
+  [
+    "ceil",
+    "чирикГони"
+  ],
+  [
+    "cos",
+    "гопосинос"
+  ],
+  [
+    "floor",
+    "бабкиГони"
+  ],
+  [
+    "round",
+    "подрезать"
+  ],
+  [
+    "tan",
+    "гопангенс"
+  ],
+  [
+    "this",
+    "тырыпыры"
+  ],
+  [
+    "while",
+    "потрещим"
+  ],
+  [
+    "import",
+    "спиздить"
+  ],
+  [
+    "protected",
+    "подкрыша"
+  ],
+  [
+    "false",
+    "нетрулио"
+  ],
+  [
+    "++",
+    "плюсуюНа"
+  ],
+  [
+    "--",
+    "слилсяНа"
+  ],
+  [
+    "evaluate",
+    "заценить"
+  ],
+  [
+    "doctype",
+    "типКсивы"
+  ],
+  [
+    "anchors",
+    "якоряЁпт"
+  ],
+  [
+    "anchors",
+    "якоряЕпт"
+  ],
+  [
+    "dir",
+    "буратино"
+  ],
+  [
+    "location",
+    "райончик"
+  ],
+  [
+    "onfocus",
+    "опаЧотка"
+  ],
+  [
+    "onclick",
+    "какПырну"
+  ],
+  [
+    "onclose",
+    "ебалоОфф"
+  ],
+  [
+    "onplay",
+    "покаЖиви"
+  ],
+  [
+    "onsort",
+    "опаСидор"
+  ],
+  [
+    "closed",
+    "завалено"
+  ],
+  [
+    "name",
+    "погоняло"
+  ],
+  [
+    "screen",
+    "всёЕбало"
+  ],
+  [
+    "screen",
+    "всеЕбало"
+  ],
+  [
+    "self",
+    "пельмень"
+  ],
+  [
+    "find",
+    "сигиЕсть"
+  ],
+  [
+    "print",
+    "наПечать"
+  ],
+  [
+    "scroll",
+    "колесить"
+  ],
+  [
+    "onstorage",
+    "опаХабар"
+  ],
+  [
+    "parentNode",
+    "братишка"
+  ],
+  [
+    "normalize",
+    "нормандэ"
+  ],
+  [
+    "indexOf",
+    "поТюряге"
+  ],
+  [
+    "toLowerCase",
+    "неКапсом"
+  ],
+  [
+    "anchor",
+    "якорьЁпт"
+  ],
+  [
+    "anchor",
+    "якорьЕпт"
+  ],
+  [
+    "filter",
+    "фильтруй"
+  ],
+  [
+    "arcTo",
+    "прогибНа"
+  ],
+  [
+    "clip",
+    "запретка"
+  ],
+  [
+    "moveTo",
+    "щемитьНа"
+  ],
+  [
+    "isInteger",
+    "этоХуйня"
+  ],
+  [
+    "isNaN",
+    "этоНихуя"
+  ],
+  [
+    "error",
+    "папандос"
+  ],
+  [
+    "of",
+    "сашаГрей"
+  ],
+  [
+    "asin",
+    "агопинус"
+  ],
+  [
+    "debugger",
+    "логопед"
+  ],
+  [
+    "default",
+    "понятие"
+  ],
+  [
+    "finally",
+    "тюряжка"
+  ],
+  [
+    "instanceof",
+    "внатуре"
+  ],
+  [
+    "new",
+    "гыйбать"
+  ],
+  [
+    "return",
+    "отвечаю"
+  ],
+  [
+    "try",
+    "хапнуть"
+  ],
+  [
+    "Const",
+    "ясенХуй"
+  ],
+  [
+    "Double",
+    "двойные"
+  ],
+  [
+    "export",
+    "излабас"
+  ],
+  [
+    "Float",
+    "плавник"
+  ],
+  [
+    "implements",
+    "силикон"
+  ],
+  [
+    "Long",
+    "колонна"
+  ],
+  [
+    "package",
+    "клеёнка"
+  ],
+  [
+    "package",
+    "клеенка"
+  ],
+  [
+    "public",
+    "ебанное"
+  ],
+  [
+    "throws",
+    "плюнуть"
+  ],
+  [
+    "==",
+    "эквалио"
+  ],
+  [
+    "charset",
+    "слышЁба"
+  ],
+  [
+    "charset",
+    "слышЕба"
+  ],
+  [
+    "timeline",
+    "всяЖиза"
+  ],
+  [
+    "children",
+    "пездюки"
+  ],
+  [
+    "embeds",
+    "мразоты"
+  ],
+  [
+    "head",
+    "залупка"
+  ],
+  [
+    "height",
+    "длинный"
+  ],
+  [
+    "plugins",
+    "выебоны"
+  ],
+  [
+    "scripts",
+    "гыебаты"
+  ],
+  [
+    "onshow",
+    "опаТуса"
+  ],
+  [
+    "dump",
+    "мусорка"
+  ],
+  [
+    "valueOf",
+    "валиоОф"
+  ],
+  [
+    "small",
+    "малорик"
+  ],
+  [
+    "scale",
+    "чотчеНа"
+  ],
+  [
+    "stroke",
+    "наколка"
+  ],
+  [
+    "Array",
+    "Помойка"
+  ],
+  [
+    "keys",
+    "отмычки"
+  ],
+  [
+    "reduce",
+    "редиска"
+  ],
+  [
+    "values",
+    "валиоси"
+  ],
+  [
+    "sin",
+    "гопинус"
+  ],
+  [
+    "continue",
+    "двигай"
+  ],
+  [
+    "delete",
+    "ёбнуть"
+  ],
+  [
+    "delete",
+    "ебнуть"
+  ],
+  [
+    "with",
+    "хзйопт"
+  ],
+  [
+    "synchronized",
+    "вписон"
+  ],
+  [
+    "volatile",
+    "вписос"
+  ],
+  [
+    "null",
+    "нуллио"
+  ],
+  [
+    "true",
+    "трулио"
+  ],
+  [
+    "<",
+    "хуёвей"
+  ],
+  [
+    "<",
+    "хуевей"
+  ],
+  [
+    "close",
+    "завали"
+  ],
+  [
+    "open",
+    "отрыть"
+  ],
+  [
+    "write",
+    "малява"
+  ],
+  [
+    "hidden",
+    "кроить"
+  ],
+  [
+    "forms",
+    "еблища"
+  ],
+  [
+    "width",
+    "жирный"
+  ],
+  [
+    "controllers",
+    "мусора"
+  ],
+  [
+    "parent",
+    "родаки"
+  ],
+  [
+    "confirm",
+    "канает"
+  ],
+  [
+    "focus",
+    "хуёкус"
+  ],
+  [
+    "focus",
+    "хуекус"
+  ],
+  [
+    "stop",
+    "стопээ"
+  ],
+  [
+    "raw",
+    "полоса"
+  ],
+  [
+    "charAt",
+    "обаЁба"
+  ],
+  [
+    "charAt",
+    "обаЕба"
+  ],
+  [
+    "match",
+    "футбик"
+  ],
+  [
+    "direction",
+    "лесТам"
+  ],
+  [
+    "arc",
+    "прогиб"
+  ],
+  [
+    "table",
+    "таблом"
+  ],
+  [
+    "every",
+    "естьЧо"
+  ],
+  [
+    "PI",
+    "ПИЗДЕЦ"
+  ],
+  [
+    "min",
+    "хуйчик"
+  ],
+  [
+    "pow",
+    "гопень"
+  ],
+  [
+    "sign",
+    "сиськи"
+  ],
+  [
+    "sqrt",
+    "сквирт"
+  ],
+  [
+    "function",
+    "йопта"
+  ],
+  [
+    "Boolean",
+    "пацан"
+  ],
+  [
+    "Byte",
+    "семка"
+  ],
+  [
+    "Class",
+    "клёво"
+  ],
+  [
+    "Class",
+    "клево"
+  ],
+  [
+    "final",
+    "бачок"
+  ],
+  [
+    "goto",
+    "ходка"
+  ],
+  [
+    "Int",
+    "хуйня"
+  ],
+  [
+    "interface",
+    "хуёво"
+  ],
+  [
+    "interface",
+    "хуево"
+  ],
+  [
+    "Short",
+    "пипин"
+  ],
+  [
+    "NaN",
+    "нихуя"
+  ],
+  [
+    ">=",
+    "поцик"
+  ],
+  [
+    "||",
+    "иличо"
+  ],
+  [
+    ">",
+    "пизже"
+  ],
+  [
+    "!",
+    "чобля"
+  ],
+  [
+    "document",
+    "ксива"
+  ],
+  [
+    "clear",
+    "урыть"
+  ],
+  [
+    "body",
+    "висяк"
+  ],
+  [
+    "cookie",
+    "семки"
+  ],
+  [
+    "images",
+    "мазни"
+  ],
+  [
+    "value",
+    "валио"
+  ],
+  [
+    "window",
+    "ебало"
+  ],
+  [
+    "frames",
+    "кадры"
+  ],
+  [
+    "alert",
+    "шухер"
+  ],
+  [
+    "rect",
+    "доска"
+  ],
+  [
+    "toFixed",
+    "наПМС"
+  ],
+  [
+    "group",
+    "банда"
+  ],
+  [
+    "profile",
+    "личка"
+  ],
+  [
+    "pop",
+    "попка"
+  ],
+  [
+    "push",
+    "пупок"
+  ],
+  [
+    "sort",
+    "сидор"
+  ],
+  [
+    "Math",
+    "Гопец"
+  ],
+  [
+    "max",
+    "хуйло"
+  ],
+  [
+    "break",
+    "харэ"
+  ],
+  [
+    "void",
+    "куку"
+  ],
+  [
+    "extends",
+    "батя"
+  ],
+  [
+    "native",
+    "чорт"
+  ],
+  [
+    "=",
+    "сука"
+  ],
+  [
+    ";",
+    " нах"
+  ],
+  [
+    "async",
+    "ассо"
+  ],
+  [
+    "links",
+    "зоны"
+  ],
+  [
+    "title",
+    "вася"
+  ],
+  [
+    "URL",
+    "хата"
+  ],
+  [
+    "link",
+    "зона"
+  ],
+  [
+    "ellipse",
+    "очко"
+  ],
+  [
+    "info",
+    "инфо"
+  ],
+  [
+    "time",
+    "срок"
+  ],
+  [
+    "random",
+    "шара"
+  ],
+  [
+    "case",
+    "лещ"
+  ],
+  [
+    "catch",
+    "гоп"
+  ],
+  [
+    "do",
+    "крч"
+  ],
+  [
+    "throw",
+    "пнх"
+  ],
+  [
+    "var",
+    "гыы"
+  ],
+  [
+    "Char",
+    "эээ"
+  ],
+  [
+    "Enum",
+    "еээ"
+  ],
+  [
+    "private",
+    "мой"
+  ],
+  [
+    "super",
+    "яга"
+  ],
+  [
+    "transient",
+    "ахз"
+  ],
+  [
+    "<=",
+    "поц"
+  ],
+  [
+    "&&",
+    "ичо"
+  ],
+  [
+    "history",
+    "фон"
+  ],
+  [
+    "fixed",
+    "ПМС"
+  ],
+  [
+    "log",
+    "чмо"
+  ],
+  [
+    "for",
+    "го"
+  ],
+  [
+    "in",
+    "в"
+  ]
+];
