--- src/yoptascript.js
@@ -1,2672 +1,0 @@
-/*!
- * YoptaScript v0.2.7 (https://yopta.space)
- * Copyright (c) 2016 Yopta.Space project and Contributors
- * Licensed under the MIT license
- */
-(function (w, d) {
-    /* YoptaScript
-     * Массив без лишних экранирований
-     * Для перевода из js в ys
-     */
-    var dictionary = [
-        [
-            "includes",
-            "лучшеНетВлагалищаЧемОчкоТоварища"
-        ],
-        [
-            "createRadialGradient",
-            "намутитьПоКругуМазнюПодливой"
-        ],
-        [
-            "createLinearGradient",
-            "намутитьЧоткуюМазнюПодливой"
-        ],
-        [
-            "resizeTo",
-            "распидораситьОтносительно"
-        ],
-        [
-            "fromCodePoint",
-            "хуйняИзЭтоТуданахНутыпоэл"
-        ],
-        [
-            "createTextNode",
-            "намутитьМалявуГовнодскую"
-        ],
-        [
-            "enableStyleSheetsForSet",
-            "намутитьСтруйкуДляХабара"
-        ],
-        [
-            "ondragenter",
-            "покаДёргалкаНарисовалась"
-        ],
-        [
-            "ondragenter",
-            "покаДергалкаНарисовалась"
-        ],
-        [
-            "imul",
-            "петухПетухаВидитИздалека"
-        ],
-        [
-            "getElementsByClassName",
-            "вычислитьТерпилПоКлассу"
-        ],
-        [
-            "vlinkColor",
-            "когдаОткинулсяПослеЗоны"
-        ],
-        [
-            "onpointermove",
-            "покаХожуТырколйПоРайону"
-        ],
-        [
-            "getElementsByName",
-            "вычислитьТерпилПоИмени"
-        ],
-        [
-            "onpointerenter",
-            "покаСрелкаНарисовалась"
-        ],
-        [
-            "oninstall",
-            "покаХуйСтоитКакКолонна"
-        ],
-        [
-            "isPointInStroke",
-            "естьЛиКуполаНаНаколках"
-        ],
-        [
-            "getElementsByTagName",
-            "вычислитьТерпилПоТегу"
-        ],
-        [
-            "getElementById",
-            "вычислитьЛохаПоНомеру"
-        ],
-        [
-            "querySelectorAll",
-            "хулиТутВсеВыёбываются"
-        ],
-        [
-            "querySelectorAll",
-            "хулиТутВсеВыебываются"
-        ],
-        [
-            "pointerLockElement",
-            "тырколкуНаАнусТерпилы"
-        ],
-        [
-            "getComputedStyle",
-            "нассыМнеВалиоСтруйкой"
-        ],
-        [
-            "imageSmoothingEnabled",
-            "включитьРазмытиеЕбала"
-        ],
-        [
-            "scrollPathIntoView",
-            "колеситьПоДорогеНахуй"
-        ],
-        [
-            "setLineDash",
-            "заебенитьПятнистыйХуй"
-        ],
-        [
-            "onreadystatechange",
-            "опаГотовЙоптЧозанахуй"
-        ],
-        [
-            "resolve",
-            "щащаНамутитьКактоНадо"
-        ],
-        [
-            "createNodeIterator",
-            "намутитьГовнодыратор"
-        ],
-        [
-            "lastElementChild",
-            "последнийПездюкШняги"
-        ],
-        [
-            "onafterscriptexecute",
-            "послеВыполненияЙопты"
-        ],
-        [
-            "scrollbars",
-            "колеситьПоПивларькам"
-        ],
-        [
-            "resizeBy",
-            "распидораситьПоХуйне"
-        ],
-        [
-            "sizeToContent",
-            "ухтыжёптыжбляПодгони"
-        ],
-        [
-            "sizeToContent",
-            "ухтыжептыжбляПодгони"
-        ],
-        [
-            "fontcolor",
-            "говномПоСтенеКрасиво"
-        ],
-        [
-            "removeHitRegion",
-            "избавитьсяОтПроблемы"
-        ],
-        [
-            "getResponseHeader",
-            "дайКепарикПолосатого"
-        ],
-        [
-            "acosh",
-            "агопосинусКупчинский"
-        ],
-        [
-            "atanh",
-            "агопангенсКупчинский"
-        ],
-        [
-            "createDocumentFragment",
-            "намутитьКусокМалявы"
-        ],
-        [
-            "onselectionchange",
-            "покаДваСтулаМахнуть"
-        ],
-        [
-            "onfullscreenchange",
-            "покаЕбальникПоказал"
-        ],
-        [
-            "ondblclick",
-            "какПырнуДваждыНахуй"
-        ],
-        [
-            "ondragend",
-            "покаХарэОчкоДёргать"
-        ],
-        [
-            "ondragend",
-            "покаХарэОчкоДергать"
-        ],
-        [
-            "oninvalid",
-            "гыйбатьИнвалидНахуй"
-        ],
-        [
-            "onmouseout",
-            "покаТырколкаСъебала"
-        ],
-        [
-            "onmouseover",
-            "покаТырколкаПодошла"
-        ],
-        [
-            "onpointercancel",
-            "покаСтрелаТруханула"
-        ],
-        [
-            "onpopstate",
-            "покаИсторияМаляется"
-        ],
-        [
-            "isDefaultNamespace",
-            "деткаТыПростоКосмос"
-        ],
-        [
-            "codePointAt",
-            "хуйняНутыпоэлОткуда"
-        ],
-        [
-            "isPointInPath",
-            "естьЛиКуполаНаГруди"
-        ],
-        [
-            "XMLHttpRequestEventTarget",
-            "запросСоШнягойПоЩам"
-        ],
-        [
-            "cosh",
-            "гопосиносКолпинский"
-        ],
-        [
-            "tanh",
-            "гопангенсКолпинский"
-        ],
-        [
-            "getOwnPropertyNames",
-            "вычислиПогонялаКрыс"
-        ],
-        [
-            "setPrototypeOf",
-            "замутитьПроточелика"
-        ],
-        [
-            "ondrop",
-            "опаМабилкаЁбнулась"
-        ],
-        [
-            "ondrop",
-            "опаМабилкаЕбнулась"
-        ],
-        [
-            "oninput",
-            "покаЭйтыэтоПишибля"
-        ],
-        [
-            "ontouchmove",
-            "опаДвигайОтСюдаЛещ"
-        ],
-        [
-            "scrollMaxX",
-            "колеситьПоГлавЭксу"
-        ],
-        [
-            "scrollMaxY",
-            "колеситьПоГлавУгам"
-        ],
-        [
-            "compareDocumentPosition",
-            "сравниСтатусМалявы"
-        ],
-        [
-            "substring",
-            "спиздитьМеждуСтрок"
-        ],
-        [
-            "trimRight",
-            "вырезатьОчкоСправа"
-        ],
-        [
-            "fontsize",
-            "говномПоСтенеСочно"
-        ],
-        [
-            "createImageData",
-            "намутитьМазнюЙопта"
-        ],
-        [
-            "setTransform",
-            "перекоситьЕбальник"
-        ],
-        [
-            "parseFloat",
-            "шнырятьПоПлавникам"
-        ],
-        [
-            "dirxml",
-            "йбатьБуратиноНахуй"
-        ],
-        [
-            "readystatechange",
-            "готовЙоптЧозанахуй"
-        ],
-        [
-            "LOG10E",
-            "СЛОЖНЫЙ_ГОПОРИФМ10"
-        ],
-        [
-            "SQRT1_2",
-            "сквиртНаПолшишечки"
-        ],
-        [
-            "asinh",
-            "агопинусКупчинский"
-        ],
-        [
-            "queryCommandEnabled",
-            "хулиЧикаДоступная"
-        ],
-        [
-            "queryCommandSupported",
-            "хулиЧикаБезАйфона"
-        ],
-        [
-            "firstElementChild",
-            "первыйПездюкШняги"
-        ],
-        [
-            "childElementCount",
-            "моиШняжныеПездюки"
-        ],
-        [
-            "onbeforescriptexecute",
-            "доВыполненияЙопты"
-        ],
-        [
-            "ondragover",
-            "покаДёрнулПодошла"
-        ],
-        [
-            "ondragover",
-            "покаДернулПодошла"
-        ],
-        [
-            "onpointerover",
-            "покаСтрелаПодошла"
-        ],
-        [
-            "onpointerout",
-            "покаСтрелаСъебала"
-        ],
-        [
-            "onpointerleave",
-            "покаСтрелаСлиняла"
-        ],
-        [
-            "ontouchstart",
-            "опаЩаЛещаПоЩамДам"
-        ],
-        [
-            "dispatchEvent",
-            "послатьНахуйШнягу"
-        ],
-        [
-            "scrollByPages",
-            "колеситьНаМалявах"
-        ],
-        [
-            "updateCommands",
-            "новыйАйфонДляЧики"
-        ],
-        [
-            "onbeforeunload",
-            "покаСемкиКрутятся"
-        ],
-        [
-            "onhashchange",
-            "покаШнягаИзменяет"
-        ],
-        [
-            "onpagehide",
-            "покаКсиваНаМалине"
-        ],
-        [
-            "hasChildNodes",
-            "батяИмеетПездюков"
-        ],
-        [
-            "insertBefore",
-            "вставитьПездюкаДо"
-        ],
-        [
-            "substr",
-            "спиздитьМеждуБукв"
-        ],
-        [
-            "trimLeft",
-            "вырезатьОчкоСлева"
-        ],
-        [
-            "quadraticCurveTo",
-            "криваяЗавелаНахуй"
-        ],
-        [
-            "resetTransform",
-            "поправитьЕбальник"
-        ],
-        [
-            "parseInt",
-            "шнырятьПоКарманам"
-        ],
-        [
-            "timeout",
-            "длительностьСрока"
-        ],
-        [
-            "loadstart",
-            "началТаскатьСемки"
-        ],
-        [
-            "progress",
-            "сколькоСемокДонёс"
-        ],
-        [
-            "progress",
-            "сколькоСемокДонес"
-        ],
-        [
-            "LOG2E",
-            "СЛОЖНЫЙ_ГОПОРИФМ2"
-        ],
-        [
-            "clz32",
-            "поводырь32петухов"
-        ],
-        [
-            "log1p",
-            "чистыйГопорифмПо1"
-        ],
-        [
-            "sinh",
-            "гопинусКолпинский"
-        ],
-        [
-            "getOwnPropertyDescriptors",
-            "вычислиСходкуКрыс"
-        ],
-        [
-            "getOwnPropertySymbols",
-            "выучиАлфавитМразь"
-        ],
-        [
-            "createTreeWalker",
-            "намутитьБуратино"
-        ],
-        [
-            "getAnimations",
-            "вычислитьДвижуху"
-        ],
-        [
-            "releaseCapture",
-            "зафотатьХуякХуяк"
-        ],
-        [
-            "querySelector",
-            "хулиВыёбываешься"
-        ],
-        [
-            "querySelector",
-            "хулиВыебываешься"
-        ],
-        [
-            "queryCommandState",
-            "хулиЧикаОтдыхает"
-        ],
-        [
-            "contentType",
-            "ухтыжёптыжТипчик"
-        ],
-        [
-            "contentType",
-            "ухтыжептыжТипчик"
-        ],
-        [
-            "alinkColor",
-            "петухЗоныКрасиво"
-        ],
-        [
-            "lastModified",
-            "когдаПетухомСтал"
-        ],
-        [
-            "ondragexit",
-            "анусСебеДёрниПёс"
-        ],
-        [
-            "ondragexit",
-            "анусСебеДерниПес"
-        ],
-        [
-            "ondragleave",
-            "покаДергунСлинял"
-        ],
-        [
-            "onmouseleave",
-            "съебалсяИзРайона"
-        ],
-        [
-            "onpointerdown",
-            "покаТыркнулДауна"
-        ],
-        [
-            "onprogress",
-            "покаМатаетсяСрок"
-        ],
-        [
-            "ontimeupdate",
-            "опаНуЭтоКогдаЭто"
-        ],
-        [
-            "addEventListener",
-            "добавитьВертухай"
-        ],
-        [
-            "setResizable",
-            "датьПопидорасить"
-        ],
-        [
-            "onunload",
-            "опаСемкиКрутятся"
-        ],
-        [
-            "baseURIObject",
-            "мразотыНаХатеКрч"
-        ],
-        [
-            "ownerDocument",
-            "главныйАвторитет"
-        ],
-        [
-            "textContent",
-            "ухтыжёптыжМалява"
-        ],
-        [
-            "textContent",
-            "ухтыжептыжМалява"
-        ],
-        [
-            "contains",
-            "яТвойОтецЕбуОвец"
-        ],
-        [
-            "lastIndexOf",
-            "последняяОтсидка"
-        ],
-        [
-            "toLocaleLowerCase",
-            "поРайонуНеКапсом"
-        ],
-        [
-            "shadowBlur",
-            "наняРазмытьЕбало"
-        ],
-        [
-            "isSafeInteger",
-            "этоОхуеннаяХуйня"
-        ],
-        [
-            "toLocaleString",
-            "кПацанамНаРайоне"
-        ],
-        [
-            "groupEnd",
-            "съебатьсяИзБанды"
-        ],
-        [
-            "reverse",
-            "шухильмеМухильме"
-        ],
-        [
-            "resolveAfter2Seconds",
-            "паруСекНеГомосек"
-        ],
-        [
-            "defineProperties",
-            "ващеЧоткиеЧелики"
-        ],
-        [
-            "getOwnPropertyDescriptor",
-            "вычислиЕблоКрысы"
-        ],
-        [
-            "isSealed",
-            "теЧоЕбалоРазбить"
-        ],
-        [
-            "elementFromPoint",
-            "терпилаИзПараши"
-        ],
-        [
-            "elementsFromPoint",
-            "терпилыИзПараши"
-        ],
-        [
-            "getSelection",
-            "сестьНаДваСтула"
-        ],
-        [
-            "onabort",
-            "когдаУронилМыло"
-        ],
-        [
-            "onkeypress",
-            "гыйбатьВмялЛеща"
-        ],
-        [
-            "onpointerup",
-            "покаТыркнулАута"
-        ],
-        [
-            "ontouchcancel",
-            "покаЛещТруханул"
-        ],
-        [
-            "returnValue",
-            "ответитьЗаВалио"
-        ],
-        [
-            "scrollByLines",
-            "колеситьНаЛинии"
-        ],
-        [
-            "scrollTo",
-            "колеситьНахуйНа"
-        ],
-        [
-            "onpageshow",
-            "опаКсивуПоказал"
-        ],
-        [
-            "lastChild",
-            "последнийПездюк"
-        ],
-        [
-            "nextSibling",
-            "следующийПездюк"
-        ],
-        [
-            "nodeName",
-            "погонялоПездюка"
-        ],
-        [
-            "parentElement",
-            "братишкаЭлемент"
-        ],
-        [
-            "appendChild",
-            "заделатьПездюка"
-        ],
-        [
-            "fillStyle",
-            "кончитьСтруйкой"
-        ],
-        [
-            "miterLimit",
-            "скрестимСтруйки"
-        ],
-        [
-            "addHitRegion",
-            "создатьПроблемы"
-        ],
-        [
-            "createPattern",
-            "намутитьТипчика"
-        ],
-        [
-            "getImageData",
-            "чоТутНамалёвано"
-        ],
-        [
-            "getImageData",
-            "чоТутНамалевано"
-        ],
-        [
-            "profileEnd",
-            "вЛичкуПрописали"
-        ],
-        [
-            "constructor",
-            "стрельнутьБычки"
-        ],
-        [
-            "'use strict'",
-            "\"далиСтрогача\""
-        ],
-        [
-            "onwheel",
-            "какНаХуюВертел"
-        ],
-        [
-            "ondragstart",
-            "покаДёрнулКмон"
-        ],
-        [
-            "ondragstart",
-            "покаДернулКмон"
-        ],
-        [
-            "onkeydown",
-            "гыйбатьЛещДаун"
-        ],
-        [
-            "onmousedown",
-            "всунулНаРайоне"
-        ],
-        [
-            "onmousewheel",
-            "вертелНаРайоне"
-        ],
-        [
-            "onseeking",
-            "когдаОбоссался"
-        ],
-        [
-            "onstalled",
-            "опаНефартануло"
-        ],
-        [
-            "screenX",
-            "всёЕбалоПоЭксу"
-        ],
-        [
-            "screenX",
-            "всеЕбалоПоЭксу"
-        ],
-        [
-            "screenY",
-            "всёЕбалоПоУгам"
-        ],
-        [
-            "screenY",
-            "всеЕбалоПоУгам"
-        ],
-        [
-            "scrollX",
-            "колеситьПоЭксу"
-        ],
-        [
-            "scrollY",
-            "колеситьПоУгам"
-        ],
-        [
-            "ondevicelight",
-            "покаХуйДлинный"
-        ],
-        [
-            "onpaint",
-            "опаНарисовался"
-        ],
-        [
-            "nodeType",
-            "типичныйПездюк"
-        ],
-        [
-            "padStart",
-            "начальныйПадик"
-        ],
-        [
-            "toLocaleUpperCase",
-            "поРайонуКапсом"
-        ],
-        [
-            "shadowOffsetX",
-            "наняХатаПоЭксу"
-        ],
-        [
-            "shadowOffsetY",
-            "наняХатаПоУгам"
-        ],
-        [
-            "textAlign",
-            "выровнитьБазар"
-        ],
-        [
-            "textBaseline",
-            "поставитьБазар"
-        ],
-        [
-            "fillRect",
-            "обкончатьДоску"
-        ],
-        [
-            "fillText",
-            "обкончатьБуквы"
-        ],
-        [
-            "MAX_VALUE",
-            "ОХУЕННОЕ_ВАЛИО"
-        ],
-        [
-            "toExponential",
-            "наХуекспоненту"
-        ],
-        [
-            "trace",
-            "складЧмошников"
-        ],
-        [
-            "readyState",
-            "готовностьЙопт"
-        ],
-        [
-            "entries",
-            "вычислитьЛохов"
-        ],
-        [
-            "expm1",
-            "топГопспонента"
-        ],
-        [
-            "fround",
-            "мелочьТожеГони"
-        ],
-        [
-            "ignoreCase",
-            "игнорщикЕбаный"
-        ],
-        [
-            "sticky",
-            "петухОпущенный"
-        ],
-        [
-            "lastIndex",
-            "доКонцаОтсидки"
-        ],
-        [
-            "test",
-            "ответыБудутЭээ"
-        ],
-        [
-            "getPrototypeOf",
-            "чоЗаПроточелик"
-        ],
-        [
-            "captureEvents",
-            "зафотатьШняги"
-        ],
-        [
-            "createEvent",
-            "намутитьШнягу"
-        ],
-        [
-            "createRange",
-            "намутитьОпапа"
-        ],
-        [
-            "createElement",
-            "намутитьЛошка"
-        ],
-        [
-            "createExpression",
-            "намутитьБазар"
-        ],
-        [
-            "execCommand",
-            "идиРаботайБля"
-        ],
-        [
-            "hasFocus",
-            "имеетЧеткость"
-        ],
-        [
-            "queryCommandIndeterm",
-            "хулиЧикаОйВсё"
-        ],
-        [
-            "queryCommandIndeterm",
-            "хулиЧикаОйВсе"
-        ],
-        [
-            "queryCommandValue",
-            "хулиЧикаВалио"
-        ],
-        [
-            "inputEncoding",
-            "эйтыэтоПиздиш"
-        ],
-        [
-            "scrollingElement",
-            "намазиТерпила"
-        ],
-        [
-            "activeElement",
-            "активнаяШняга"
-        ],
-        [
-            "defaultView",
-            "моргалаВыколю"
-        ],
-        [
-            "ondrag",
-            "опаОчкоДёрнул"
-        ],
-        [
-            "ondrag",
-            "опаОчкоДернул"
-        ],
-        [
-            "onkeyup",
-            "гыйбатьЛещАут"
-        ],
-        [
-            "onloadstart",
-            "покаНесуСемки"
-        ],
-        [
-            "onmouseup",
-            "вынулНаРайоне"
-        ],
-        [
-            "onsuspend",
-            "опаПодфартило"
-        ],
-        [
-            "ontouchend",
-            "покаЛещКончил"
-        ],
-        [
-            "onwaiting",
-            "покаМотаюСрок"
-        ],
-        [
-            "frameElement",
-            "кадрОпущенный"
-        ],
-        [
-            "innerHeight",
-            "внутриДлинный"
-        ],
-        [
-            "outerHeight",
-            "вокругДлинные"
-        ],
-        [
-            "sessionStorage",
-            "хабрИзОтсидки"
-        ],
-        [
-            "postMessage",
-            "намутитьКсиву"
-        ],
-        [
-            "prompt",
-            "поясниЗаБазар"
-        ],
-        [
-            "removeEventListener",
-            "урытьВертухая"
-        ],
-        [
-            "getRootNode",
-            "дайБатеГовнод"
-        ],
-        [
-            "isEqualNode",
-            "эквалиоГовнод"
-        ],
-        [
-            "removeChild",
-            "уебатьПездюка"
-        ],
-        [
-            "localeCompare",
-            "сравнитьГовор"
-        ],
-        [
-            "padEnd",
-            "залупныйПадик"
-        ],
-        [
-            "repeat",
-            "непоэлПовтори"
-        ],
-        [
-            "slice",
-            "поделитьСемки"
-        ],
-        [
-            "startsWith",
-            "начатьЗалупку"
-        ],
-        [
-            "font",
-            "говномПоСтене"
-        ],
-        [
-            "globalAlpha",
-            "главныйАльфач"
-        ],
-        [
-            "clearHitRegions",
-            "убитьПроблему"
-        ],
-        [
-            "measureText",
-            "вместитьБазар"
-        ],
-        [
-            "restore",
-            "выздоравливай"
-        ],
-        [
-            "MAX_SAFE_INTEGER",
-            "ЛУЧШИЙ_ГАНДОН"
-        ],
-        [
-            "MIN_SAFE_INTEGER",
-            "ХУЁВЫЙ_ГАНДОН"
-        ],
-        [
-            "MIN_SAFE_INTEGER",
-            "ХУЕВЫЙ_ГАНДОН"
-        ],
-        [
-            "groupCollapsed",
-            "свернутьБанду"
-        ],
-        [
-            "EventTarget",
-            "очкоНаПрицеле"
-        ],
-        [
-            "reduceRight",
-            "редискаПравая"
-        ],
-        [
-            "SQRT2",
-            "двойнойСквирт"
-        ],
-        [
-            "atan2",
-            "агопангенсПо2"
-        ],
-        [
-            "hypot",
-            "вКореньЗыришь"
-        ],
-        [
-            "RegExp",
-            "фильтруйБазар"
-        ],
-        [
-            "global",
-            "глобалкаЙопта"
-        ],
-        [
-            "assign",
-            "тащиВсёНаХату"
-        ],
-        [
-            "assign",
-            "тащиВсеНаХату"
-        ],
-        [
-            "defineProperty",
-            "вотЭтоЗаебись"
-        ],
-        [
-            "isPrototypeOf",
-            "чейПроточелик"
-        ],
-        [
-            "createAttribute",
-            "намутитьЯжку"
-        ],
-        [
-            "createTouch",
-            "намутитьЛеща"
-        ],
-        [
-            "documentElement",
-            "ксиваТерпилы"
-        ],
-        [
-            "bgColor",
-            "охуеннаяЖопа"
-        ],
-        [
-            "oncancel",
-            "покаТруханул"
-        ],
-        [
-            "onchange",
-            "опаЧозанахуй"
-        ],
-        [
-            "oncontextmenu",
-            "какПоЛбуЁбну"
-        ],
-        [
-            "oncontextmenu",
-            "какПоЛбуЕбну"
-        ],
-        [
-            "onmouseenter",
-            "вошёлНаРайон"
-        ],
-        [
-            "onmouseenter",
-            "вошелНаРайон"
-        ],
-        [
-            "onmousemove",
-            "хожуПоРайону"
-        ],
-        [
-            "onseeked",
-            "когдаОбоссал"
-        ],
-        [
-            "onsubmit",
-            "опаХуйВГовне"
-        ],
-        [
-            "onvolumechange",
-            "покаТишеБудь"
-        ],
-        [
-            "console",
-            "красноглазое"
-        ],
-        [
-            "dialogArguments",
-            "тыэтаТавоэта"
-        ],
-        [
-            "innerWidth",
-            "внутриЖирный"
-        ],
-        [
-            "outerWidth",
-            "вокругЖирные"
-        ],
-        [
-            "pageXOffset",
-            "статьяПоЭксу"
-        ],
-        [
-            "pageYOffset",
-            "статьяПоУгам"
-        ],
-        [
-            "blur",
-            "размытьЕбало"
-        ],
-        [
-            "clearInterval",
-            "отсидетьСизо"
-        ],
-        [
-            "clearTimeout",
-            "отсидетьСрок"
-        ],
-        [
-            "moveTo",
-            "нахуйЭтоТуда"
-        ],
-        [
-            "setInterval",
-            "посетитьСизо"
-        ],
-        [
-            "setTimeout",
-            "получитьСрок"
-        ],
-        [
-            "onload",
-            "опаСемкиНесу"
-        ],
-        [
-            "childNodes",
-            "пездюкГовнод"
-        ],
-        [
-            "firstChild",
-            "первыйПездюк"
-        ],
-        [
-            "nodeValue",
-            "валиоПездюка"
-        ],
-        [
-            "previousSibling",
-            "старыйПездюк"
-        ],
-        [
-            "replaceChild",
-            "сделатьАборт"
-        ],
-        [
-            "replace",
-            "пивасПодмени"
-        ],
-        [
-            "trim",
-            "вырезатьОчко"
-        ],
-        [
-            "currentTransform",
-            "этотЕбальник"
-        ],
-        [
-            "lineCap",
-            "выбратьКонец"
-        ],
-        [
-            "lineDashOffset",
-            "пятнистыйХуй"
-        ],
-        [
-            "lineJoin",
-            "формаШишкана"
-        ],
-        [
-            "strokeStyle",
-            "стильНаколок"
-        ],
-        [
-            "putImageData",
-            "намалюйЧоТут"
-        ],
-        [
-            "rotate",
-            "вертетьНаХую"
-        ],
-        [
-            "strokeText",
-            "текстНаколки"
-        ],
-        [
-            "translate",
-            "дисюдаПиксел"
-        ],
-        [
-            "MIN_VALUE",
-            "ХУЁВОЕ_ВАЛИО"
-        ],
-        [
-            "MIN_VALUE",
-            "ХУЕВОЕ_ВАЛИО"
-        ],
-        [
-            "POSITIVE_INFINITY",
-            "ОХУЕТЬ_ДОХУЯ"
-        ],
-        [
-            "assert",
-            "найтиСтукача"
-        ],
-        [
-            "copyWithin",
-            "вынестиГовно"
-        ],
-        [
-            "forEach",
-            "пероПодРебро"
-        ],
-        [
-            "splice",
-            "въебатьГовна"
-        ],
-        [
-            "log10",
-            "гопорифмПо10"
-        ],
-        [
-            "input",
-            "тыЭтоПишибля"
-        ],
-        [
-            "flags",
-            "флагМнеВанус"
-        ],
-        [
-            "exec",
-            "работайМразь"
-        ],
-        [
-            "hasOwnProperty",
-            "соСвоейТемой"
-        ],
-        [
-            "else",
-            "иливжопураз"
-        ],
-        [
-            "if",
-            "вилкойвглаз"
-        ],
-        [
-            "characterSet",
-            "слышТыЧоЁба"
-        ],
-        [
-            "characterSet",
-            "слышТыЧоЕба"
-        ],
-        [
-            "documentURI",
-            "ксиваНаХате"
-        ],
-        [
-            "domConfig",
-            "чоткоДерзко"
-        ],
-        [
-            "visibilityState",
-            "мутныйСюжет"
-        ],
-        [
-            "designMode",
-            "хуйРисуйМод"
-        ],
-        [
-            "linkColor",
-            "зонаКрасиво"
-        ],
-        [
-            "referrer",
-            "корешСтарый"
-        ],
-        [
-            "onscroll",
-            "покаКолесим"
-        ],
-        [
-            "onselect",
-            "опаДваСтула"
-        ],
-        [
-            "location",
-            "белыйЛебедь"
-        ],
-        [
-            "sidebar",
-            "стенкаЙбать"
-        ],
-        [
-            "cancelIdleCallback",
-            "харэПиздеть"
-        ],
-        [
-            "maximize",
-            "распидорась"
-        ],
-        [
-            "moveBy",
-            "щаТяПодвину"
-        ],
-        [
-            "Promise",
-            "СловоПацана"
-        ],
-        [
-            "onoffline",
-            "покаОффнусь"
-        ],
-        [
-            "charCodeAt",
-            "обаЁбаХуйня"
-        ],
-        [
-            "charCodeAt",
-            "обаЕбаХуйня"
-        ],
-        [
-            "split",
-            "поделитьЯгу"
-        ],
-        [
-            "italics",
-            "понаехавший"
-        ],
-        [
-            "shadowColor",
-            "наняКрасиво"
-        ],
-        [
-            "getContext",
-            "снятьСкальп"
-        ],
-        [
-            "NEGATIVE_INFINITY",
-            "НИХУЯ_ДОХУЯ"
-        ],
-        [
-            "isFinite",
-            "оноКонченое"
-        ],
-        [
-            "count",
-            "которыйСрок"
-        ],
-        [
-            "timeStamp",
-            "началоСрока"
-        ],
-        [
-            "XMLHttpRequest",
-            "запросПоЩам"
-        ],
-        [
-            "shift",
-            "первыйБачок"
-        ],
-        [
-            "some",
-            "нарываешься"
-        ],
-        [
-            "E",
-            "ГОПСПАНЕНТА"
-        ],
-        [
-            "exp",
-            "гопспанента"
-        ],
-        [
-            "log2",
-            "гопорифмПо2"
-        ],
-        [
-            "await",
-            "сидетьНахуй"
-        ],
-        [
-            "all",
-            "пацанСделал"
-        ],
-        [
-            "race",
-            "пацанСказал"
-        ],
-        [
-            "isExtensible",
-            "жратьБудешь"
-        ],
-        [
-            "seal",
-            "сдохниНахуй"
-        ],
-        [
-            "exports",
-            "предъявляет"
-        ],
-        [
-            "let",
-            "участковый"
-        ],
-        [
-            "Abstract",
-            "говнойбать"
-        ],
-        [
-            "importNode",
-            "влабазУзел"
-        ],
-        [
-            "registerElement",
-            "зашитьДело"
-        ],
-        [
-            "domain",
-            "домойБлядь"
-        ],
-        [
-            "oncut",
-            "какВысунул"
-        ],
-        [
-            "onpaste",
-            "какВставил"
-        ],
-        [
-            "onblur",
-            "опаНичотка"
-        ],
-        [
-            "onerror",
-            "наПапандос"
-        ],
-        [
-            "onreset",
-            "покаПравим"
-        ],
-        [
-            "fullScreen",
-            "воВсёЕбало"
-        ],
-        [
-            "fullScreen",
-            "воВсеЕбало"
-        ],
-        [
-            "opener",
-            "открывашка"
-        ],
-        [
-            "performance",
-            "сестьНахуй"
-        ],
-        [
-            "top",
-            "КрышаЙбать"
-        ],
-        [
-            "getAttention",
-            "посвистеть"
-        ],
-        [
-            "matchMedia",
-            "феняНаШару"
-        ],
-        [
-            "openDialog",
-            "побазарить"
-        ],
-        [
-            "scrollBy",
-            "колеситьНа"
-        ],
-        [
-            "ononline",
-            "опаТутачки"
-        ],
-        [
-            "cloneNode",
-            "клонГовнод"
-        ],
-        [
-            "prototype",
-            "проточелик"
-        ],
-        [
-            "fromCharCode",
-            "хуйняИзЁба"
-        ],
-        [
-            "fromCharCode",
-            "хуйняИзЕба"
-        ],
-        [
-            "concat",
-            "заебеньВсе"
-        ],
-        [
-            "toString",
-            "поПацански"
-        ],
-        [
-            "big",
-            "большойЁпт"
-        ],
-        [
-            "big",
-            "большойЕпт"
-        ],
-        [
-            "lineWidth",
-            "толщинаХуя"
-        ],
-        [
-            "bezierCurveTo",
-            "кривоНахуй"
-        ],
-        [
-            "closePath",
-            "тупикНахуй"
-        ],
-        [
-            "getLineDash",
-            "сококПятен"
-        ],
-        [
-            "lineTo",
-            "прямоНахуй"
-        ],
-        [
-            "transform",
-            "перекосить"
-        ],
-        [
-            "timeEnd",
-            "конецСрока"
-        ],
-        [
-            "warn",
-            "тыЭтоНуЭто"
-        ],
-        [
-            "withCredentials",
-            "зашкварить"
-        ],
-        [
-            "abort",
-            "уронилМыло"
-        ],
-        [
-            "load",
-            "нестиСемки"
-        ],
-        [
-            "loadend",
-            "семкиДонёс"
-        ],
-        [
-            "loadend",
-            "семкиДонес"
-        ],
-        [
-            "from",
-            "спиздитьИз"
-        ],
-        [
-            "isArray",
-            "этоПомойка"
-        ],
-        [
-            "findIndex",
-            "найдиБомжа"
-        ],
-        [
-            "map",
-            "засратьВсё"
-        ],
-        [
-            "map",
-            "засратьВсе"
-        ],
-        [
-            "unshift",
-            "верниБачок"
-        ],
-        [
-            "LN10",
-            "ГОПОРИФМ10"
-        ],
-        [
-            "acos",
-            "агопосинус"
-        ],
-        [
-            "atan",
-            "агопангенс"
-        ],
-        [
-            "cbrt",
-            "кубоСквирт"
-        ],
-        [
-            "trunc",
-            "верниЧирик"
-        ],
-        [
-            "multiline",
-            "стулБезПик"
-        ],
-        [
-            "unicode",
-            "хуйняНахуй"
-        ],
-        [
-            "reject",
-            "пацанЗабыл"
-        ],
-        [
-            "watch",
-            "смотрюСюда"
-        ],
-        [
-            "switch",
-            "пошерстим"
-        ],
-        [
-            "typeof",
-            "чезажижан"
-        ],
-        [
-            "static",
-            "попонятия"
-        ],
-        [
-            "==",
-            "однахуйня"
-        ],
-        [
-            "writeln",
-            "малявагоп"
-        ],
-        [
-            "readyState",
-            "газуемБля"
-        ],
-        [
-            "oncopy",
-            "какВсунул"
-        ],
-        [
-            "onpause",
-            "покаСтопэ"
-        ],
-        [
-            "onplaying",
-            "покаЖивой"
-        ],
-        [
-            "crypto",
-            "пиздишбля"
-        ],
-        [
-            "devicePixelRatio",
-            "типАйфона"
-        ],
-        [
-            "length",
-            "писькомер"
-        ],
-        [
-            "navigator",
-            "главпетух"
-        ],
-        [
-            "minimize",
-            "спидорась"
-        ],
-        [
-            "baseURI",
-            "наХатеТип"
-        ],
-        [
-            "endsWith",
-            "отЗалупки"
-        ],
-        [
-            "search",
-            "семкиЕсть"
-        ],
-        [
-            "toUpperCase",
-            "капсомБля"
-        ],
-        [
-            "bold",
-            "жирныйЁпт"
-        ],
-        [
-            "bold",
-            "жирныйЕпт"
-        ],
-        [
-            "strike",
-            "въебиОчко"
-        ],
-        [
-            "clearRect",
-            "урытьШкаф"
-        ],
-        [
-            "fill",
-            "обкончать"
-        ],
-        [
-            "save",
-            "схоронить"
-        ],
-        [
-            "EPSILON",
-            "ХУЕПСИЛОН"
-        ],
-        [
-            "toPrecision",
-            "кЧоткости"
-        ],
-        [
-            "upload",
-            "заебенить"
-        ],
-        [
-            "send",
-            "всёПиздуй"
-        ],
-        [
-            "send",
-            "всеПиздуй"
-        ],
-        [
-            "join",
-            "вписаться"
-        ],
-        [
-            "LN2",
-            "ГОПОРИФМ2"
-        ],
-        [
-            "abs",
-            "абсолютли"
-        ],
-        [
-            "ceil",
-            "чирикГони"
-        ],
-        [
-            "cos",
-            "гопосинос"
-        ],
-        [
-            "floor",
-            "бабкиГони"
-        ],
-        [
-            "round",
-            "подрезать"
-        ],
-        [
-            "tan",
-            "гопангенс"
-        ],
-        [
-            "AsyncFunction",
-            "АссоЙопта"
-        ],
-        [
-            "propertyIsEnumerable",
-            "лютаяТема"
-        ],
-        [
-            "this",
-            "тырыпыры"
-        ],
-        [
-            "while",
-            "потрещим"
-        ],
-        [
-            "import",
-            "спиздить"
-        ],
-        [
-            "protected",
-            "подкрыша"
-        ],
-        [
-            "false",
-            "нетрулио"
-        ],
-        [
-            "++",
-            "плюсуюНа"
-        ],
-        [
-            "--",
-            "слилсяНа"
-        ],
-        [
-            "evaluate",
-            "заценить"
-        ],
-        [
-            "doctype",
-            "типКсивы"
-        ],
-        [
-            "anchors",
-            "якоряЁпт"
-        ],
-        [
-            "anchors",
-            "якоряЕпт"
-        ],
-        [
-            "dir",
-            "буратино"
-        ],
-        [
-            "location",
-            "райончик"
-        ],
-        [
-            "onfocus",
-            "опаЧотка"
-        ],
-        [
-            "onclick",
-            "какПырну"
-        ],
-        [
-            "onclose",
-            "ебалоОфф"
-        ],
-        [
-            "onplay",
-            "покаЖиви"
-        ],
-        [
-            "onsort",
-            "опаСидор"
-        ],
-        [
-            "closed",
-            "завалено"
-        ],
-        [
-            "name",
-            "погоняло"
-        ],
-        [
-            "screen",
-            "всёЕбало"
-        ],
-        [
-            "screen",
-            "всеЕбало"
-        ],
-        [
-            "self",
-            "пельмень"
-        ],
-        [
-            "find",
-            "сигиЕсть"
-        ],
-        [
-            "print",
-            "наПечать"
-        ],
-        [
-            "scroll",
-            "колесить"
-        ],
-        [
-            "onstorage",
-            "опаХабар"
-        ],
-        [
-            "parentNode",
-            "братишка"
-        ],
-        [
-            "normalize",
-            "нормандэ"
-        ],
-        [
-            "indexOf",
-            "поТюряге"
-        ],
-        [
-            "toLowerCase",
-            "неКапсом"
-        ],
-        [
-            "anchor",
-            "якорьЁпт"
-        ],
-        [
-            "anchor",
-            "якорьЕпт"
-        ],
-        [
-            "filter",
-            "фильтруй"
-        ],
-        [
-            "arcTo",
-            "прогибНа"
-        ],
-        [
-            "clip",
-            "запретка"
-        ],
-        [
-            "moveTo",
-            "щемитьНа"
-        ],
-        [
-            "isInteger",
-            "этоХуйня"
-        ],
-        [
-            "isNaN",
-            "этоНихуя"
-        ],
-        [
-            "error",
-            "папандос"
-        ],
-        [
-            "of",
-            "сашаГрей"
-        ],
-        [
-            "asin",
-            "агопинус"
-        ],
-        [
-            "create",
-            "намутить"
-        ],
-        [
-            "freeze",
-            "датьЛеща"
-        ],
-        [
-            "export",
-            "предъявa"
-        ],
-        [
-            "debugger",
-            "логопед"
-        ],
-        [
-            "finally",
-            "тюряжка"
-        ],
-        [
-            "instanceof",
-            "внатуре"
-        ],
-        [
-            "new",
-            "гыйбать"
-        ],
-        [
-            "return",
-            "отвечаю"
-        ],
-        [
-            "try",
-            "хапнуть"
-        ],
-        [
-            "Const",
-            "ясенХуй"
-        ],
-        [
-            "Double",
-            "двойные"
-        ],
-        [
-            "Float",
-            "плавник"
-        ],
-        [
-            "implements",
-            "силикон"
-        ],
-        [
-            "Long",
-            "колонна"
-        ],
-        [
-            "package",
-            "клеёнка"
-        ],
-        [
-            "package",
-            "клеенка"
-        ],
-        [
-            "public",
-            "ебанное"
-        ],
-        [
-            "throws",
-            "плюнуть"
-        ],
-        [
-            "false",
-            "пиздишь"
-        ],
-        [
-            "==",
-            "эквалио"
-        ],
-        [
-            "===",
-            "блябуду"
-        ],
-        [
-            "charset",
-            "слышЁба"
-        ],
-        [
-            "charset",
-            "слышЕба"
-        ],
-        [
-            "timeline",
-            "всяЖиза"
-        ],
-        [
-            "children",
-            "пездюки"
-        ],
-        [
-            "embeds",
-            "мразоты"
-        ],
-        [
-            "head",
-            "залупка"
-        ],
-        [
-            "height",
-            "длинный"
-        ],
-        [
-            "plugins",
-            "выебоны"
-        ],
-        [
-            "scripts",
-            "гыебаты"
-        ],
-        [
-            "onshow",
-            "опаТуса"
-        ],
-        [
-            "dump",
-            "мусорка"
-        ],
-        [
-            "valueOf",
-            "валиоОф"
-        ],
-        [
-            "small",
-            "малорик"
-        ],
-        [
-            "scale",
-            "чотчеНа"
-        ],
-        [
-            "stroke",
-            "наколка"
-        ],
-        [
-            "Array",
-            "Помойка"
-        ],
-        [
-            "keys",
-            "отмычки"
-        ],
-        [
-            "reduce",
-            "редиска"
-        ],
-        [
-            "values",
-            "валиоси"
-        ],
-        [
-            "sin",
-            "гопинус"
-        ],
-        [
-            "source",
-            "обоснуй"
-        ],
-        [
-            "isFrozen",
-            "далЛеща"
-        ],
-        [
-            "unwatch",
-            "съебись"
-        ],
-        [
-            "continue",
-            "двигай"
-        ],
-        [
-            "delete",
-            "ёбнуть"
-        ],
-        [
-            "delete",
-            "ебнуть"
-        ],
-        [
-            "with",
-            "хзйопт"
-        ],
-        [
-            "synchronized",
-            "вписон"
-        ],
-        [
-            "volatile",
-            "вписос"
-        ],
-        [
-            "null",
-            "нуллио"
-        ],
-        [
-            "true",
-            "трулио"
-        ],
-        [
-            "<",
-            "хуёвей"
-        ],
-        [
-            "<",
-            "хуевей"
-        ],
-        [
-            "close",
-            "завали"
-        ],
-        [
-            "open",
-            "отрыть"
-        ],
-        [
-            "write",
-            "малява"
-        ],
-        [
-            "hidden",
-            "кроить"
-        ],
-        [
-            "forms",
-            "еблища"
-        ],
-        [
-            "width",
-            "жирный"
-        ],
-        [
-            "controllers",
-            "мусора"
-        ],
-        [
-            "parent",
-            "родаки"
-        ],
-        [
-            "confirm",
-            "канает"
-        ],
-        [
-            "focus",
-            "хуёкус"
-        ],
-        [
-            "focus",
-            "хуекус"
-        ],
-        [
-            "stop",
-            "стопээ"
-        ],
-        [
-            "raw",
-            "полоса"
-        ],
-        [
-            "charAt",
-            "обаЁба"
-        ],
-        [
-            "charAt",
-            "обаЕба"
-        ],
-        [
-            "match",
-            "футбик"
-        ],
-        [
-            "direction",
-            "лесТам"
-        ],
-        [
-            "arc",
-            "прогиб"
-        ],
-        [
-            "table",
-            "таблом"
-        ],
-        [
-            "every",
-            "естьЧо"
-        ],
-        [
-            "PI",
-            "ПИЗДЕЦ"
-        ],
-        [
-            "min",
-            "хуйчик"
-        ],
-        [
-            "pow",
-            "гопень"
-        ],
-        [
-            "sign",
-            "сиськи"
-        ],
-        [
-            "sqrt",
-            "сквирт"
-        ],
-        [
-            "then",
-            "атоэто"
-        ],
-        [
-            "module",
-            "братва"
-        ],
-        [
-            "default",
-            "пахану"
-        ],
-        [
-            "function",
-            "йопта"
-        ],
-        [
-            "Boolean",
-            "пацан"
-        ],
-        [
-            "Byte",
-            "семка"
-        ],
-        [
-            "class",
-            "клёво"
-        ],
-        [
-            "class",
-            "клево"
-        ],
-        [
-            "final",
-            "бачок"
-        ],
-        [
-            "goto",
-            "ходка"
-        ],
-        [
-            "Int",
-            "хуйня"
-        ],
-        [
-            "interface",
-            "хуёво"
-        ],
-        [
-            "interface",
-            "хуево"
-        ],
-        [
-            "Short",
-            "пипин"
-        ],
-        [
-            "NaN",
-            "нихуя"
-        ],
-        [
-            "true",
-            "чётко"
-        ],
-        [
-            "true",
-            "четко"
-        ],
-        [
-            "==",
-            "ровно"
-        ],
-        [
-            ">=",
-            "поцик"
-        ],
-        [
-            "||",
-            "иличо"
-        ],
-        [
-            ">",
-            "пизже"
-        ],
-        [
-            "!",
-            "чобля"
-        ],
-        [
-            "document",
-            "ксива"
-        ],
-        [
-            "clear",
-            "урыть"
-        ],
-        [
-            "body",
-            "висяк"
-        ],
-        [
-            "cookie",
-            "семки"
-        ],
-        [
-            "images",
-            "мазни"
-        ],
-        [
-            "value",
-            "валио"
-        ],
-        [
-            "window",
-            "ебало"
-        ],
-        [
-            "frames",
-            "кадры"
-        ],
-        [
-            "alert",
-            "шухер"
-        ],
-        [
-            "rect",
-            "доска"
-        ],
-        [
-            "toFixed",
-            "наПМС"
-        ],
-        [
-            "group",
-            "банда"
-        ],
-        [
-            "profile",
-            "личка"
-        ],
-        [
-            "pop",
-            "попка"
-        ],
-        [
-            "push",
-            "пупок"
-        ],
-        [
-            "sort",
-            "сидор"
-        ],
-        [
-            "Math",
-            "Гопец"
-        ],
-        [
-            "max",
-            "хуйло"
-        ],
-        [
-            "Object",
-            "Петух"
-        ],
-        [
-            "global",
-            "общак"
-        ],
-        [
-            "break",
-            "харэ"
-        ],
-        [
-            "void",
-            "куку"
-        ],
-        [
-            "extends",
-            "батя"
-        ],
-        [
-            "native",
-            "чорт"
-        ],
-        [
-            "=",
-            "сука"
-        ],
-        [
-            ";",
-            " нах"
-        ],
-        [
-            "links",
-            "зоны"
-        ],
-        [
-            "title",
-            "вася"
-        ],
-        [
-            "URL",
-            "хата"
-        ],
-        [
-            "link",
-            "зона"
-        ],
-        [
-            "ellipse",
-            "очко"
-        ],
-        [
-            "info",
-            "инфо"
-        ],
-        [
-            "time",
-            "срок"
-        ],
-        [
-            "random",
-            "шара"
-        ],
-        [
-            "async",
-            "ассо"
-        ],
-        [
-            "case",
-            "лещ"
-        ],
-        [
-            "catch",
-            "гоп"
-        ],
-        [
-            "do",
-            "крч"
-        ],
-        [
-            "throw",
-            "пнх"
-        ],
-        [
-            "var",
-            "гыы"
-        ],
-        [
-            "Char",
-            "эээ"
-        ],
-        [
-            "Enum",
-            "еээ"
-        ],
-        [
-            "private",
-            "мой"
-        ],
-        [
-            "super",
-            "яга"
-        ],
-        [
-            "transient",
-            "ахз"
-        ],
-        [
-            "<=",
-            "поц"
-        ],
-        [
-            "&&",
-            "ичо"
-        ],
-        [
-            "history",
-            "фон"
-        ],
-        [
-            "fixed",
-            "ПМС"
-        ],
-        [
-            "log",
-            "чмо"
-        ],
-        [
-            "for",
-            "го"
-        ],
-        [
-            "in",
-            "в"
-        ]
-    ];
-
-    function escapeRegExp(str) {
-        return str.replace(/[-\/\\^$*+?.()|[\]{}]/g, '\\$&');
-    }
-
-    function yoptReplaceAll(str, search, replacement) {
-        var re = new RegExp(escapeRegExp(search), 'g');
-        return str.replace(re, replacement);
-    }
-
-    function compile(yoptaText) {
-        for (var i = 0; i < dictionary.length; i++) {
-            yoptaText = yoptReplaceAll(yoptaText, dictionary[i][1], dictionary[i][0]);
-        }
-        return yoptaText;
-    }
-
-    function yoptify(jscode) {
-        for (i = 0; i < dictionary.length; i++) {
-            jscode = yoptReplaceAll(jscode, dictionary[i][0], dictionary[i][1]);
-        }
-        return jscode;
-    }
-
-    function yoptaToJs(yopta) {
-        //Получаем йопту из скрипта
-        var yoptaText = compile(yopta.textContent);
-        if (!yoptaText.length) {
-            //Пошли искать сорцы
-            var src = yopta.getAttribute('src');
-            if (src !== null && src.length) {
-                var xml = new XMLHttpRequest();
-                xml.open('GET', src, false);
-                xml.send(null);
-                if (xml.status == 200 || xml.status == 0) yoptaText = compile(xml.responseText);
-            }
-        }
-        //удаляем старый скрипт
-        yopta.parentNode.removeChild(yopta);
-        //создаём обработанный скрипт с блекджеком и шлюхами
-        var js_script = d.createElement("script");
-        js_script.innerHTML = yoptaText;
-        document.body.appendChild(js_script);
-    }
-    //Получаем йопту из скрипта
-    d.querySelectorAll('[language="YoptaScript"]').forEach(yoptaToJs);
-    w.yopt = {
-        dictionary: dictionary,
-        compile: compile,
-        yoptify: yoptify,
-    }
-})(window, document);
